chore(couponDiscountType): 할인 유형 관련 타입을 상수 파일이 아닌 타입 파일로 이동
