+++ src/constants/couponDiscountType.ts
@@ -4,13 +4,3 @@ export const COUPON_DISCOUNT_TYPE = {
   buyXgetY: "buyXgetY",
   freeShipping: "freeShipping",
 } as const;
-
-export type CouponDiscountType = FixedDiscount | PercentageDiscount | BuyXGetY | FreeShipping;
-
-export type FixedDiscount = typeof COUPON_DISCOUNT_TYPE.fixed;
-
-export type PercentageDiscount = typeof COUPON_DISCOUNT_TYPE.percentage;
-
-export type BuyXGetY = typeof COUPON_DISCOUNT_TYPE.buyXgetY;
-
-export type FreeShipping = typeof COUPON_DISCOUNT_TYPE.freeShipping;
